+++ src/app/locales/en/components/qos-parameters-suggestion-box.js
@@ -1,0 +1,6 @@
+export default {
+  key: 'Key',
+  value: 'Value',
+  insert: 'Insert',
+  custom: 'custom',
+};